chore: ability to expire feature flags
Some feature flags in CDKv1 will be entirely removed in CDKv2, and
their 'enabled' behaviour made the default.

This change adds the ability to expire these flags. The list currently
stands empty and will stand empty in the `master` branch. It will be
populated with feature flags in the `v2-main` branch.