chore: ability to expire feature flags (#12688)
Some feature flags in CDKv1 will be entirely removed in CDKv2, and
their 'enabled' behaviour made the default.

This change adds the ability to expire these flags. The list currently
stands empty and will stand empty in the `master` branch. It will be
populated with feature flags in the `v2-main` branch.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*