chore(aws-cdk-lib): revert raising the min node version to v14
This reverts commit 4aa64c82b6849b06ee6d18ad040d5dbaef6d8732

This change is incompatible with our build image 'jsii/superchain' that
has node v10 as its requirement. We would still like the default version
of this build image to stay at v10 since that is the minimum supported
version of CDKv1.

Additional work is required in this area before we can raise the node
version requirement for this package.