docs(route53): Add example of A record for Elastic IP (#12176)
It took me ages to figure out how to add an A record for an EIP.

Using `instance.instancePublicIp` doesn't work, because it seems the instance is given another public IP before the EIP is associated with it, and that is what `instance.instancePublicIp` resolves to, which then quickly becomes wrong.

It's not clear in `RecordTarget.fromIpAddresses` that it can take a reference to an EIP. I'm still not even sure why this works, but it does. So I thought I'd add an example to help others see how to do it.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*