fix(pipelines): unable to publish assets inside VPC (#12331)
The CodeBuild projects for the asset publishing could not be
created inside a VPC, as the Exeuction Roles assigned to it
were missing the required permissions.

Normally the CodeBuild project would take care of these permissions,
but since we are creating a single role that is shared between all
CodeBuild projects (in order to not explode the number of roles in the
account) and the role we manage is immutable, the pipeline library
itself is responsible for adding the permissions.

Fixes #11815


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*