chore(cdk-build-tools): temporarily disable feature flag test (#12711)
This was originally introduced in 31b1b3289b99c4ec4ed19c51705d66f0e83783dd.

However, the forward merge of this change onto the v2-main branch is
failing unit tests in consuming modules that implicitly
depend on the future flag's behaviour. 
For example, unit tests in the `aws-s3` module depend on the
behaviour of the `@aws-cdk/aws-kms:defaultKeyPolicies` flag.

Temporarily disable this feature until these can be resolved.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*