feat(core): manually control exports to solve "deadly embrace"
"Deadly embrace" is a great term but an annoying problem. The best
way to work around it is to manually ensure the CloudFormation Export
exists while you remove the consuming relationship. Adam has a blog
post about this, but the mechanism can be more smooth.

Add a method, `stack.exportAttribute(...)` which can be used to
create the Export for the duration of the deployment that breaks
the relationship, and add an explanation of how to use it.

Fixes #7602.