fix(codedeploy): allow the install agent script's commands to exit with errors
The script that installed the CodeDeploy agent on the hosts in the ServerDeploymentGroup relies on some commands failing,
and checking their return codes.
Unfortunately, calling the `addExecuteFileCommand()` method on the ASG sets the 'set -e' flag,
which makes the script fail.

Explicitly call the 'set +e' command in the script as the first instruction.

Fixes #12764