fix(efs): EFS fails to create when using a VPC with multiple subnets per availability zone (#12097)
Fixes #10170

### Fixes

This PR has been modified so that it does not create a mount target if only VPC is specified in the EFS file system. 

### Bug

Currently, if a VPC is specified and no subnet is specified, the subnet search criteria are passed as undefined, all subnets in the VPC are retrieved, and mount targets are generated for all subnets in the VPC. I will. Since mount targets can only be created in one subnet for each Availability Zone, the above behavior will result in duplicate Availability Zones between mount targets, resulting in an error when creating the mount target.



----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*