chore(eslint-plugin-cdk): fix tests and expectations (#12831)
The previous change - 9a81faaa - to add test cases to this package
had a bug. Two different eslint rules were being applied
simultaneously creating corrupt expectation.

Fixing the test so it only runs the specific eslint rule for that
fixture, and fixing the expectation.

At the same time, added a test case that was previously missed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*