chore: isolate CoreConstruct imports to avoid merge conflicts (#12850)
As part of CDKv2, the construct layer in '@aws-cdk/core' is
being replaced with the constructs API in the 'constructs'
module. During the transition period between v1 and v2,
both exist.

On the `master` branch, code will include importing
`Construct` from the '@aws-cdk/core' module but this
will not be possible in the `v2-main` branch where this API
does not exist.

Any change to the same line as the import of the `Construct`
symbol will cause merge conflicts. Hence, move the import
into a separate line and section to avoid these conflicts.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*