chore: fix bug introduced in recent PR
In #12765, a bug was introduced where there was a mismatch between
the file entry written to the asset manifest for the stack template,
and the template location written to the cloud assembly, thereby
causing all deployments to fail.

The error was "Access Denied" (because that's what S3 says when it
doesn't want to tell you "File not found").

Fix the inconsistency.