chore: run pkglint via Github actions for v2 pull requests
The [merge-forward.sh] script runs a 'git merge' followed by pkglint in
automated scheduled job, which finally posts a pull request with the
changes.

If any merge conflicts arise during the run which need manual
intervention, the job simply fails and does not post the pull request.
This is sub-optimal. It would be a much nicer experience if the pull
request was posted along with the merge conflicts, allowing users to
resolve them on the pull request using the Github web experience.

This commit changes the flow so that pkglint is run as a Github action
after the pull request is posted, thereby enabling this experience.

[merge-forward.sh]: https://github.com/aws/aws-cdk/blob/b4e1b682604946c30ab63164017c73bd8d828728/scripts/merge-forward.sh