feat(ec2): can define Launch Templates (not use them yet) (#12385)
This provides an initial implementation of a level 2 construct for EC2 Launch Templates. It is a start that is intended to help get the ball rolling on implementation of Launch Template support within the CDK. It is a step towards resolving https://github.com/aws/aws-cdk/issues/6734

Launch Templates have value even without the integrations into Instance and AutoScalingGroup being implemented yet. Thus, the intention with this PR is to solely implement the L2 for LaunchTemplate. Future work in a separate PR would be required to implement its integration into Instance, AutoScalingGroup, and others.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*