fix(ec2): Subnet cidr missing for Vpc.from_lookup() (#12878)
When importing a VPC using `Vpc.fromLookup()`, the subnets are properly looked up and saved into the `cdk.context.json`. However, the Subnets object within the imported VPC don't have the CIDR attached and trying to access the subnet CIDR within the code triggers the following error:

> You cannot reference an imported Subnet's IPv4 CIDR if it was not supplied. Add the ipv4CidrBlock when importing using Subnet.fromSubnetAttributes().

This PR fixes the issue.

closes #11821

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*