fix(kms): cross-environment usage fails when trustAccountIdentities is set (#12925)
When the `trustAccountIdentities` flag is set (either directly, or set by
default via the `@aws-cdk/aws-kms:defaultKeyPolicies` feature flag), any grant
statements are always added to the principal OR resource, defaulting to the
principal if possible.

In cross-environment usage, this means that the principal IAM policy is updated,
but the key policy is never updated. However, cross-environment usage always
requires the key policy to be updated. Attempting to use the key in a
cross-environment usage with `trustAccountIdentities` enabled initially presents
itself as a cross-environment reference issue, but even if the stack was
successfully deployed, the delegated account would not have access to use/admin
the key.

This change updates the logic to make use of the existing cross-environment
logic whenever cross-environment usage is detected, regardless of the value of
`trustAccountIdentities`. This has the impact of fixing the cross-env references
and ensuring the key policy is properly updated.

fixes #12921, fixes #12741


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*