chore: start token identifiers at a random number
It's far too easy to write:

```ts
JSON.stringify({ ...struct that can contain stringified tokens... })
```

Which should have been:

```ts
JSON.stringify(stack.resolve({ ...struct... }))
```

This typically manifests when hashing some structure, and in the hash
it's hard to see that wrong components went into calculating it.

Make sure the stringified token values are different every run, so
that it becomes impossible to write a test that (consistently) succeeds
if it's hashing unresolved tokens.