fix(cfn2ts): correctly choose between string and object without required properties in a union
In our code generation, in the validation functions for CFN complex types,
we never checked whether the argument we received is an object.
Because of that, passing a string was actually a correct candidate for a union with a complex type that had no required properties.
Fix that by checking for that error explicitly,
which removes the complex type without required properties from the candidates when a string is passed.

Fixes #12854