chore: prepare tests to expire feature flag 'kms:defaultKeyPolicies' (#12966)
This is a follow up to - 31b1b32.

Unit tests in the `aws-ec2` and `aws-s3-notifications` modules
assert resources generated by KMS.

The '@aws-cdk/aws-kms:defaultKeyPolicies' feature flag will be
reset in CDKv2. Update the necessary tests so that they continue
to work when this happens.

Two course of actions are taken here - some tests needed to
be made more 'precise' in their assertions, and others had to be
configured to only run in CDKv1.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*