fix(cfn2ts): correctly choose between string and object without required properties in a union (#12954)
In our code generation, in the validation functions for CFN complex types,
we never checked whether the argument we received is an object.
Because of that, passing a string was actually a correct candidate for a union with a complex type that had no required properties.
Fix that by checking for that error explicitly,
which removes the complex type without required properties from the candidates for the union type when a string is passed.

Fixes #12854

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*