feat: future flags 'core:enableStackNameDuplicates', 'aws-secretsmanager:parseOwnedSecretName' and 'aws-kms:defaultKeyPolicies' are no longer supported (#12644)
The following feature flags have been marked as expired -

* @aws-cdk/core:enableStackNameDuplicates
* @aws-cdk/aws-secretsmanager:parseOwnedSecretName
* @aws-cdk/aws-kms:defaultKeyPolicies

These flags will be unavailable in CDKv2 and setting this flag in the context
will result in an error.

Most of the changes here are related to the feature flag
'@aws-cdk/aws-kms:defaultKeyPolicies' which changes the KMS
permissions. Since KMS is a fundamental AWS service, used by almost all
downstream services, there is a significant change to all snapshots.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*