fix(apigatewayv2): HttpApi and Route in different stacks creates cycles (#13010)
When the Route construct with an integration is defined
in a separate stack from the one defining the HttpApi
construct, a circular dependency error is thrown.

The bug was introduced in - 855ce590 - where the
refactor of moving the creation of Integration resulted in
using the HttpApi as the scope of the Integration
construct, rather than the Route.

fixes #13021

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*