fix(cloudfront): EdgeFunction us-east-1 stack created in different account (#13055)
The us-east-1 stack for EdgeFunction was defaulting to the default deploy
account. This means that using one account to deploy, and another for the stack
which contains the EdgeFunction, the support stack in us-east-1 will be deployed
into the deploy account, rather than stack account.

This fix has the us-east-1 stack inherit the account from the parent stack.

fixes #12789


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*