chore(cdk-build-tools): backport feature flag based conditional from v2-main (#13056)
This was originally commented out - d7e028a1 - since this was not yet
compatible with the v2-main branch.

Since then, the v2-main branch has been updated - 7554246ab - with the
necessary fixes and the conditional has been enabled there[1].

No effective change to master branch because of this. It reduces divergence
of the two active branches.

[1]: https://github.com/aws/aws-cdk/commit/7554246ab6d0819bc57d5e67cd9e4e10a3b7e742#diff-d325e066c7cdf5d2a8423e65b0190045defe711b4dc25adc0a94a11eef163ed9


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*