chore: enable feature flags on all integ tests (#13110)
Most feature flags will be expired in CDKv2 and their 'enabled'
behaviour made the default.
To keep the divergence between the two branches (`master` and
`v2-main`) to a minimum, update all integration tests to have the
flags enabled.

If the flag needs to be different for any specific test, it can be
overridden when constructing the `App` object.

To keep the set of changes small, only the following flags have
been added. The rest will be added in a subsequent change.
- @aws-cdk/aws-kms:defaultKeyPolicies
- @aws-cdk/core:enableStackNameDuplicates
- @aws-cdk/aws-secretsmanager:parseOwnedSecretName

Currently, only these three feature flags are expired in CDKv2.
Fixing them first will make the branch divergence smaller and
reduce already occuring [merge conflicts].

[merge conflicts]: https://github.com/aws/aws-cdk/pull/13099

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*