chore: make NPM dist tag explicit for all packages (#13109)
Currently, our NPM tag strategy is uniform across all packages. This _mostly_
works, but for V2 it makes sense for `aws-cdk-lib` to be published with a dist
tag of 'latest' while still publishing all of the other libraries as 'next'.

This change is the first supporting change to enable us to have different dist
tags for different packages.
* Introduces a new pkglint rule that enforces that each package.json has a
  publishConfig and tag. For v1, this defaults to 'latest'; for v2 it will
  default to 'next'. These are the correct values, except for `aws-cdk-lib`,
  which will need to be manually changed to `latest`.
* Removes the disttag element from `release.json`.
* Changing `check-api-compatibility` to use the dist tags from  the individual
  `package.json` files, rather than from `release.json`.

Files to look at (all others are package.json changes):
* release.json
* scripts/check-api-compatibility.sh
* scripts/resolve-version-lib.js
* scripts/script-tests/resolve-version.test.js
* tools/pkglint/lib/rules.ts

Once this is approved, and merged up to V2, the change to flip the aws-cdk-lib
tag can be made. Finally, we will update our publishing infrastructure so these
new tags are respected.

related #13058


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*