chore: enable feature flags on all integ tests (#13133)
This is the final part of the work started out in 200d52ef.

Most feature flags will be expired in CDKv2 and their 'enabled'
behaviour made the default.
To keep the divergence between the two branches (`master` and
`v2-main`) to a minimum, update all integration tests to have the
flags enabled.

If the flag needs to be different for any specific test, it can be
overridden when constructing the `App` object.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*